Start to work on porting examples to standard lib
