+++ lib/qcor_common
@@ -1,0 +1,2 @@
+#pragma once
+#include "impl/common.hpp"
